Refined aprgbooleanalgebra module

+++ AprgBooleanAlgebra/AprgBooleanAlgebra/src/BooleanAlgebra/Satisfiability/Utilities.cpp
@@ -19,15 +19,13 @@ unsigned int getSatisfiabilityLevel(
     unsigned int result(0);
     for(SatisfiabilityTerm const& satTerm : satTerms)
     {
-        result = max(result, satTerm.size());
+        result = max(result, static_cast<unsigned int>(satTerm.size()));
     }
     return result;
 }
-
 SatisfiabilityTerms getSatisfiabilityTerms(
         Term const& term)
-{
-    SatisfiabilityTerms result;
+{    SatisfiabilityTerms result;
     if(term.isVariableTerm())
     {
         result.emplace_back(SatisfiabilityTerm{term.getVariableTermConstReference()});
