+++ AprgBooleanAlgebra/AprgBooleanAlgebra/src/BooleanAlgebra/Simplification/SimplificationUtilities.cpp
@@ -23,15 +23,13 @@ namespace booleanAlgebra
 namespace Simplification
 {
 
-using Minterm = unsigned long long;
+using Minterm = uint64_t;
 using Implicant = Implicant<Minterm>;
 using Implicants = Implicants<Minterm>;
 using QuineMcCluskey = QuineMcCluskey<Minterm>;
-
 // utilties functions for this file
 namespace
-{
-Implicants getBestFinalImplicantsUsingQuineMcCluskey(
+{Implicants getBestFinalImplicantsUsingQuineMcCluskey(
         Term const& term,
         VariableNamesSet const& variableNames)
 {
